docs: 폴더위치 변경 및 주석 처리[FE-19]
