+++ features/diagnosis/ui/SymptomsList.tsx
@@ -2,17 +2,20 @@ import { FlatList, Pressable, StyleSheet } from "react-native";
 import { Label } from "@/shared/components/label/Label";
 import theme from "@/shared/styles/theme";
 import { useFormContext } from "react-hook-form"; 
-import { useSymptomsStore } from "../lib/useSymptomsStore";
+import { useSymptomsStore } from "../services/useSymptomsStore";
 import { SYMPTOMS } from '@/shared/mocks/data';
 
+// 한줄당 렌더링할 증상의 개수
 const ITEMS_PER_ROW = 5; 
 
+// 증상 카테고리 리스트 컴포넌트
 const SymptomsList = () => {
     const { setValue, watch } = useFormContext<{ symptoms: string[] }>(); 
     const selectedSymptoms: string[] = watch("symptoms") || []; 
     const { customSymptoms } = useSymptomsStore();
     const allSymptoms = [...SYMPTOMS, ...customSymptoms];
 
+    // 증상 선택시 선택된 증상 상태 변경
     const toggleSymptom = (symptom: string) => {
       const updatedSymptoms = selectedSymptoms.includes(symptom)
         ? selectedSymptoms.filter((s: string) => s !== symptom) 
@@ -21,6 +24,7 @@ const SymptomsList = () => {
       setValue("symptoms", updatedSymptoms); 
     };
 
+    // 한 줄에 몇개 씩 끊어 렌더링하기 위한 함수
     const chunkArray = (arr: string[], size: number) => {
       return Array.from({ length: Math.ceil(arr.length / size) }, (_, i) =>
         arr.slice(i * size, i * size + size)
