+++ features/diagnosis/ui/pages/StepFive.tsx
