+++ features/diagnosis/ui/pages/StepFour.tsx
