+++ features/diagnosis/ui/pages/StepOne.tsx
