+++ features/diagnosis/ui/pages/StepThree.tsx
