+++ features/diagnosis/ui/pages/StepTwo.tsx
